README.md
This project presents a machine learning-based solution for detecting and classifying different types of faults in a power distribution system, including Line-to-Ground (LG), Line-to-Line (LL), Double Line-to-Ground (LLG), and Three-Phase (3P) faults.

It uses a Kaggle dataset of simulated fault scenarios and implements models such as Random Forest and SVM in Python. The system was developed using IBM Watson Studio and deployed via IBM Cloud Lite services for real-time predictions.

The goal is to improve grid stability and operational efficiency by enabling rapid and accurate fault classification through automated analysis of voltage and current phasors.